refactor: remove `fs_cache` feature flag
