+++ Cargo.toml
@@ -83,8 +83,8 @@ json-strip-comments = "1"
 once_cell = "1" # Use `std::sync::OnceLock::get_or_try_init` when it is stable.
 papaya = "0.2.1"
 rustc-hash = { version = "2" }
-serde = { version = "1", features = ["derive"], optional = true } # derive for Deserialize from package.json
-serde_json = { version = "1", features = ["preserve_order"], optional = true } # preserve_order: package_json.exports requires order such as `["require", "import", "default"]`
+serde = { version = "1", features = ["derive"] } # derive for Deserialize from package.json
+serde_json = { version = "1", features = ["preserve_order"] } # preserve_order: package_json.exports requires order such as `["require", "import", "default"]`
 simdutf8 = { version = "0.1" }
 thiserror = "2"
 tracing = "0.1"
@@ -102,10 +102,8 @@ rayon = { version = "1.10.0" }
 vfs = "0.12.1" # for testing with in memory file system
 
 [features]
-default = ["fs_cache"]
-## Provides the `FsCache` implementation.
-fs_cache = ["dep:serde", "dep:serde_json"]
-## Enables the [PackageJsonSerde::raw_json] API,
+default = []
+## Enables the [PackageJson::raw_json] API,
 ## which returns the `package.json` with `serde_json::Value`.
 package_json_raw_json_api = []
 ## [Yarn Plug'n'Play](https://yarnpkg.com/features/pnp)
