+++ src/tests/dependencies.rs
@@ -2,10 +2,10 @@
 
 #[cfg(not(target_os = "windows"))] // MemoryFS's path separator is always `/` so the test will not pass in windows.
 mod windows {
-    use std::{path::PathBuf, sync::Arc};
+    use std::path::PathBuf;
 
     use super::super::memory_fs::MemoryFS;
-    use crate::{FsCache, ResolveContext, ResolveOptions, ResolverGeneric};
+    use crate::{ResolveContext, ResolveOptions, ResolverGeneric};
 
     fn file_system() -> MemoryFS {
         MemoryFS::new(&[
@@ -20,8 +20,8 @@ mod windows {
     fn test() {
         let file_system = file_system();
 
-        let resolver = ResolverGeneric::new_with_cache(
-            Arc::new(FsCache::new(file_system)),
+        let resolver = ResolverGeneric::new_with_file_system(
+            file_system,
             ResolveOptions {
                 extensions: vec![".json".into(), ".js".into()],
                 modules: vec!["/modules".into(), "node_modules".into()],
